Move `BinaryOperator` to the `syntax` package
